Separate utils.js to different util files
