+++ src/utils/common.js
@@ -1,10 +1,4 @@
-import dayjs from 'dayjs';
-
 const ESC_KEY_CODE = 27;
-
-export const humanizeDate = (date, formatter = `YYYY-MM-DD`) => {
-  return dayjs(date).format(formatter);
-};
 
 export const getRandomInteger = (a = 0, b = 1) => {
   const lower = Math.ceil(Math.min(a, b));
@@ -28,37 +22,6 @@ const shuffleItems = (items) => {
     shuffledItems[randomIndex] = swap;
   }
   return shuffledItems;
-};
-
-export const RenderPosition = {
-  BEFOREBEGIN: `beforebegin`,
-  AFTERBEGIN: `afterbegin`,
-  BEFOREEND: `beforeend`,
-  AFTEREND: `afterend`
-};
-
-export const render = (container, element, place) => {
-  switch (place) {
-    case RenderPosition.BEFOREBEGIN:
-      container.before(element);
-      break;
-    case RenderPosition.AFTERBEGIN:
-      container.prepend(element);
-      break;
-    case RenderPosition.BEFOREEND:
-      container.append(element);
-      break;
-    case RenderPosition.AFTEREND:
-      container.after(element);
-      break;
-  }
-};
-
-export const createElement = (template) => {
-  const newElement = document.createElement(`div`);
-  newElement.innerHTML = template;
-
-  return newElement.firstChild;
 };
 
 export const isEscEvent = (evt, action) => {
