+++ src/utils/event.js
@@ -1,0 +1,5 @@
+import dayjs from 'dayjs';
+
+export const humanizeDate = (date, formatter = `YYYY-MM-DD`) => {
+  return dayjs(date).format(formatter);
+};
